remove dangerous CheckCostWithContextIsActive rule + more descriptions and cleanup
