+++ src/test/scala/sigmastate/serialization/ErgoTreeSerializerSpecification.scala
@@ -90,7 +90,11 @@ class ErgoTreeSerializerSpecification extends SerializationSpecification
   property("max ergo tree byte size check") {
     val tree = EQ(Plus(10, 20), IntConstant(30)).toSigmaProp.treeWithSegregation
     val r = SigmaSerializer.startReader(DefaultSerializer.serializeErgoTree(tree))
-    an[InputSizeLimitExceeded] should be thrownBy DefaultSerializer.deserializeErgoTree(r, 1)
+    assertExceptionThrown({
+      DefaultSerializer.deserializeErgoTree(r, 1)
+    }, {
+      case e: SerializerException => rootCause(e).isInstanceOf[InputSizeLimitExceeded]
+    })
   }
 
   property("restore reader's positionLimit") {
